feat: console.log삭제, 폴더구조변경
