+++ src/pages/NoticeBoard.jsx
@@ -1,5 +1,5 @@
 import NavigationBar from "../../src/components/NavigationBar/NavigationBar";
-import Header from "../components/NoticeBoard/Header";
+import Header from "../components/NoticeBoard/Header/Header";
 const NoticeBoard = () => {
   return (
     <div>
