+++ src/pages/Profile.jsx
@@ -1,0 +1,11 @@
+import ProfileEdit from "../components/ProfileEdit/ProfileEdit";
+
+const Profile = () => {
+  return (
+    <div>
+      <ProfileEdit />
+    </div>
+  );
+};
+
+export default Profile;
